+++ caffe2/onnx/helper.cc
@@ -5,24 +5,27 @@
 
 namespace onnx_caffe2 {
 
-std::unordered_set<std::string> DummyName::used_names_;
 size_t DummyName::counter_ = 0;
+
+std::unordered_set<std::string>& DummyName::get_used_names() {
+  static std::unordered_set<std::string> used_names;
+  return used_names;
+}
 
 std::string DummyName::NewDummyName() {
   while (true) {
-    std::string name = caffe2::MakeString("OC2_DUMMY_", counter_++);
-    if (!used_names_.count(name)) {
-      used_names_.insert(name);
+    const std::string name = caffe2::MakeString("OC2_DUMMY_", counter_++);
+    auto ret = get_used_names().insert(name);
+    if (ret.second) {
       return name;
     }
   }
 }
 
-std::string
-DummyName::Reset(const std::unordered_set<std::string> &used_names) {
-  used_names_ = used_names;
+void DummyName::Reset(const std::unordered_set<std::string> &used_names) {
+  auto& names = get_used_names();
+  names = used_names;
   counter_ = 0;
-  return "";
 }
 
 }
